Develop

+++ src/routers/task.js
@@ -1,8 +1,13 @@
 const express = require("express");
 const Task = require("../models/task");
+const auth=require("../middleware/auth");
 const router = new express.Router();
-router.post("/tasks", async (req, res) => {
-  const task = new Task(req.body);
+router.post("/tasks",auth, async (req, res) => {
+  // const task = new Task(req.body);
+  const task= new Task({
+    ...req.body,
+    owner:req.user._id
+  })
   //   task
   //     .save()
   //     .then(() => {
@@ -18,26 +23,49 @@ router.post("/tasks", async (req, res) => {
   }
 });
 
-router.get("/tasks", async (req, res) => {
+router.get("/tasks",auth, async (req, res) => {
+  const match={};
+  const sort={};
+  if(req.query.completed)
+  {
+    match.completed=req.query.completed==='true';
+  }
+  if(req.query.sortBy)
+  {
+    const parts=req.query.sortBy.split(':');
+    sort[parts[0]]=parts[1]==='desc'?-1:1;
+  }
   // Task.find({}).then((task)=>{if(!task)return res.status(404).send;
   // res.send(task);}).catch((error) => {res.status(500).send(error)})
   try {
-    const task = await Task.find({});
-    if (!task) return res.status(404).send;
-    res.send(task);
+    // const task = await Task.find({owner: req.user._id});
+    // if (!task) return res.status(404).send;
+    await req.user.populate({
+      path:'tasks',
+      match,
+      options: {
+        limit:parseInt(req.query.limit),
+        skip:parseInt(req.query.skip),
+        sort
+      }
+    }).execPopulate();
+    res.send(req.user.tasks);
   } catch (e) {
     res.status(500).send(e);
   }
 });
-router.get("/tasks/:id", async (req, res) => {
-  const id = req.params.id;
+router.get("/tasks/:id", auth,async (req, res) => {
+  const _id = req.params.id;
   // Task.findById(id).then((task)=>{
   //     if(!task)
   //     res.status(404).send;
   //     res.send(task);
   // }).catch((error) => {res.status(500).send(error)})
   try {
-    const task = await Task.findById(id);
+    // const task = await Task.findById(id);
+    // console.log(id,req.user._id);
+    const task = await Task.findOne({ _id, owner: req.user._id })
+    console.log(task);
     if (!task) return res.status(404).send;
     res.send(task);
   } catch (e) {
@@ -45,7 +73,7 @@ router.get("/tasks/:id", async (req, res) => {
   }
 });
 
-router.patch("/tasks/:id", async (req, res) => {
+router.patch("/tasks/:id", auth,async (req, res) => {
   const updates = Object.keys(req.body);
   const allowedUpdate = ["description", "completed"];
   const isValidOperation = updates.every((update) => {
@@ -55,7 +83,7 @@ router.patch("/tasks/:id", async (req, res) => {
     return res.status(400).send({ error: "invalid updates" });
   }
   try {
-      const task=await Task.findById(req.params.id);
+    const task = await Task.findOne({ _id:req.params.id, owner: req.user._id })
       updates.forEach((update) => {
           task[update]=req.body[update];
       }) 
@@ -71,9 +99,10 @@ router.patch("/tasks/:id", async (req, res) => {
   }
 });
 
-router.delete("/tasks/:id", async (req, res) => {
+router.delete("/tasks/:id", auth,async (req, res) => {
   try {
-    const task = await Task.findOneAndDelete(req.params.id);
+    const task = await Task.findOneAndDelete({ _id:req.params.id, owner: req.user._id });
+
     if (!task) return res.status(404).send();
     res.send(task);
   } catch (e) {
